revision 3

+++ src/controllers/EventoController.js
@@ -2,9 +2,10 @@ import express from "express";
 import {EventoService} from "../servicios/EventoService.js";
 import Evento from "../entities/Evento.js";
 import AuthMiddleware from "../auth/authMiddleware.js";
-
+import {EventLocationService} from "../servicios/Event-LocationService.js"
 const router = express.Router();
 const EventService = new EventoService();
+const evLocService=new EventLocationService();
 
 
 const esFecha = (fecha) => {
@@ -40,9 +41,21 @@ router.get("/" , async (req, res) => {
 
 router.delete("/:id", AuthMiddleware , async (req, res) => {
   const id = req.params.id;
+  
   try {
-    await EventService.DeleteEvent(id);
-    return res.send("Borradisimo");4
+    const ev=await EventService.getEventById(id)
+    const tags=await EventService.getTags(id)
+    if(ev!=null){
+    if (tags!=null) {
+      
+      const eliminado=await EventService.DeleteEvent(id);
+      return res.status(200).json(eliminado)
+    }else{
+    return res.status(400).send("No se puede eliminar porque tiene tags");4
+    }
+  }else{
+    return res.status(404).send("No existe la id")
+  }
   } catch (error) {
     console.log(error);
     return res.json(error);
@@ -51,20 +64,35 @@ router.delete("/:id", AuthMiddleware , async (req, res) => {
 
 router.post("/",AuthMiddleware, async (req, res) => {
   const Evento = {};
-  Evento.name = req.query.name;
-  Evento.description = req.query.description;
-  Evento.id_event_category = req.query.id_event_category
-  Evento.id_event_location = req.query.id_event_location
-  Evento.start_date = req.query.start_date;
-  Evento.duration_in_minutes = req.query.duration_in_minutes;
-  Evento.price = req.query.price;
-  Evento.enabled_for_enrollment = req.query.enabled_for_enrollment;
-  Evento.max_assistance = req.query.max_assistance;
+  Evento.name = req.body.name;
+  Evento.description = req.body.description;
+  Evento.id_event_category = req.body.id_event_category
+  Evento.id_event_location = req.body.id_event_location
+  Evento.start_date = req.body.start_date;
+  Evento.duration_in_minutes = req.body.duration_in_minutes;
+  Evento.price = req.body.price;
+  Evento.enabled_for_enrollment = req.body.enabled_for_enrollment;
+  Evento.max_assistance = req.body.max_assistance;
   Evento.id_creator_user = req.user.id;
   
   try {
-    const respuesta = await EventService.InsertEvento(Evento);;
-    return res.json(respuesta);
+    const evLoc  =await evLocService.getEventLocationById(Evento.id_event_location)
+    if (Evento.name!=null && Evento.description!=null && Evento.id_event_category!=null && Evento.id_event_location!=null && Evento.start_date!=null && Evento.duration_in_minutes!=null && Evento.price!=null && Evento.enabled_for_enrollment!=null && Evento.max_assistance!=null) {
+      console.log(evLoc);
+      if (evLoc.max_capacity>Evento.max_assistance) {
+        if (Evento.price>0 && Evento.duration_in_minutes>0 ) {
+          const respuesta = await EventService.InsertEvento(Evento);;
+          return res.status(201).json(respuesta);
+        }else{
+          return res.status(400).send("Precio y duración no pueden ser 0")
+        }
+
+      }else{
+        return res.status(400).send("No pueden asistir más personas de las que admite la localidad")
+      }
+    }else{
+      return res.status(400).send("Faltan datos")
+    }
   } catch (error) {
     console.log(error);
     return res.json(error);
@@ -74,17 +102,42 @@ router.post("/",AuthMiddleware, async (req, res) => {
 
 router.put("/",AuthMiddleware, async (req, res) => {
   const Evento = {};
-  Evento.name = req.query.name;
-  Evento.description = req.query.description;
-  Evento.start_date = req.query.start_date;
-  Evento.duration_in_minutes = req.query.duration_in_minutes;
-  Evento.price = req.query.price;
-  Evento.enabled_for_enrollment = req.query.enabled_for_enrollment;
-  Evento.max__assistance = req.query.max__assistance;
-  Evento.id = req.query.id;
+  Evento.name = req.body.name;
+  Evento.description = req.body.description;
+  Evento.start_date = req.body.start_date;
+  Evento.duration_in_minutes = req.body.duration_in_minutes;
+  Evento.price = req.body.price;
+  Evento.enabled_for_enrollment = req.body.enabled_for_enrollment;
+  Evento.max_assistance = req.body.max_assistance;
+  Evento.id = req.body.id;
+  Evento.id_event_category=req.body.id_event_category;
+  Evento.id_event_location=req.body.id_event_location;
+    
   try {
-    const respuesta = await EventService.patchEvento(Evento);
-    return res.json(respuesta);
+    console.log(Evento);
+    const ev=await EventService.getEventById(Evento.id)
+    const evLoc  =await evLocService.getEventLocationById(Evento.id_event_location)
+    console.log(ev);
+    if (ev!=null) {
+      
+    if (Evento.name!=null && Evento.description!=null && Evento.id_event_category!=null && Evento.id_event_location!=null && Evento.start_date!=null && Evento.duration_in_minutes!=null && Evento.price!=null && Evento.enabled_for_enrollment!=null && Evento.max_assistance!=null) {
+      if (evLoc.max_capacity>Evento.max_assistance) {
+        if (Evento.price>0 && Evento.duration_in_minutes>0 ) {
+          const respuesta = await EventService.patchEvento(Evento);;
+          return res.status(201).json(respuesta);
+        }else{
+          return res.status(400).send("Precio y duración no pueden ser 0")
+        }
+
+      }else{
+        return res.status(400).send("No pueden asistir más personas de las que admite la localidad")
+      }
+    }else{
+      return res.status(400).send("Faltan datos")
+    }
+  }else{
+    return res.status(404).send("No existe un evento con esa id")
+  }
   } catch (error) {
     console.log(error);
     return res.json(error);
